Added ability to define default channel in save_account

+++ qiskit_ibm_runtime/accounts/storage.py
@@ -22,8 +22,13 @@ from .exceptions import AccountAlreadyExistsError
 logger = logging.getLogger(__name__)
 
 
-def save_config(filename: str, name: str, config: dict, overwrite: bool) -> None:
+def save_config(filename: str,
+                name: str,
+                config: dict,
+                overwrite: bool,
+                default_channel:Optional[str]=None) -> None:
     """Save configuration data in a JSON file under the given name."""
+    print("config = "+str(config))
     logger.debug("Save configuration data for '%s' in '%s'", name, filename)
     _ensure_file_exists(filename)
 
@@ -34,9 +39,15 @@ def save_config(filename: str, name: str, config: dict, overwrite: bool) -> None
         raise AccountAlreadyExistsError(
             f"Named account ({name}) already exists. " f"Set overwrite=True to overwrite."
         )
+    if data.get("default_channel") and not overwrite:
+        raise AccountAlreadyExistsError(
+            f"default_channel ({name}) already exists. " f"Set overwrite=True to overwrite."
+        )
 
     with open(filename, mode="w", encoding="utf-8") as json_out:
         data[name] = config
+        if default_channel:
+            data["default_channel"] = default_channel
         json.dump(data, json_out, sort_keys=True, indent=4)
